+++ sdk-media-webrtc/src/main/kotlin/com/pexip/sdk/media/webrtc/internal/Event.kt
@@ -17,6 +17,7 @@ package com.pexip.sdk.media.webrtc.internal
 
 import org.webrtc.IceCandidate
 import org.webrtc.PeerConnection
+import org.webrtc.PeerConnection.IceGatheringState
 import org.webrtc.RtpReceiver
 
 internal sealed interface Event {
@@ -31,6 +32,9 @@ internal sealed interface Event {
 
     @JvmInline
     value class OnIceCandidate(val candidate: IceCandidate) : Event
+
+    @JvmInline
+    value class OnIceGatheringChange(val state: IceGatheringState) : Event
 
     @JvmInline
     value class OnIceConnectionChange(val state: PeerConnection.IceConnectionState) : Event