+++ CHANGELOG.md
@@ -15,6 +15,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `Conference.referer` to improve call transfer flow
 - `MediaConnection.secureCheckCode` that can be used to verify whether a direct media connection is
   secure
+- `MediaConnectionConfig.continualGathering` to tweak ICE gathering policy
 
 ### Changed